core: refactored blockchain.go
